+++ include/zephyr/drivers/misc/timeaware_gpio/timeaware_gpio.h
@@ -75,13 +75,6 @@ __subsystem struct tgpio_driver_api {
  */
 __syscall int tgpio_port_get_time(const struct device *dev, uint64_t *current_time);
 
-static inline int z_impl_tgpio_port_get_time(const struct device *dev, uint64_t *current_time)
-{
-	const struct tgpio_driver_api *api = (const struct tgpio_driver_api *)dev->api;
-
-	return api->get_time(dev, current_time);
-}
-
 /**
  * @brief Get current running rate
  *
@@ -92,14 +85,6 @@ static inline int z_impl_tgpio_port_get_time(const struct device *dev, uint64_t 
  */
 __syscall int tgpio_port_get_cycles_per_second(const struct device *dev, uint32_t *cycles);
 
-static inline int z_impl_tgpio_port_get_cycles_per_second(const struct device *dev,
-							   uint32_t *cycles)
-{
-	const struct tgpio_driver_api *api = (const struct tgpio_driver_api *)dev->api;
-
-	return api->cyc_per_sec(dev, cycles);
-}
-
 /**
  * @brief Disable operation on pin
  *
@@ -109,13 +94,6 @@ static inline int z_impl_tgpio_port_get_cycles_per_second(const struct device *d
  * @return 0 if successful, negative errno code on failure.
  */
 __syscall int tgpio_pin_disable(const struct device *dev, uint32_t pin);
-
-static inline int z_impl_tgpio_pin_disable(const struct device *dev, uint32_t pin)
-{
-	const struct tgpio_driver_api *api = (const struct tgpio_driver_api *)dev->api;
-
-	return api->pin_disable(dev, pin);
-}
 
 /**
  * @brief Enable/Continue operation on pin
@@ -128,14 +106,6 @@ static inline int z_impl_tgpio_pin_disable(const struct device *dev, uint32_t pi
  */
 __syscall int tgpio_pin_config_ext_timestamp(const struct device *dev, uint32_t pin,
 					      uint32_t event_polarity);
-
-static inline int z_impl_tgpio_pin_config_ext_timestamp(const struct device *dev, uint32_t pin,
-							 uint32_t event_polarity)
-{
-	const struct tgpio_driver_api *api = (const struct tgpio_driver_api *)dev->api;
-
-	return api->config_ext_ts(dev, pin, event_polarity);
-}
 
 /**
  * @brief Enable periodic pulse generation on a pin
@@ -152,15 +122,6 @@ __syscall int tgpio_pin_periodic_output(const struct device *dev, uint32_t pin,
 					 uint64_t start_time, uint64_t repeat_interval,
 					 bool periodic_enable);
 
-static inline int z_impl_tgpio_pin_periodic_output(const struct device *dev, uint32_t pin,
-						    uint64_t start_time, uint64_t repeat_interval,
-						    bool periodic_enable)
-{
-	const struct tgpio_driver_api *api = (const struct tgpio_driver_api *)dev->api;
-
-	return api->set_perout(dev, pin, start_time, repeat_interval, periodic_enable);
-}
-
 /**
  * @brief Read timestamp and event counter from TGPIO
  *
@@ -174,14 +135,6 @@ static inline int z_impl_tgpio_pin_periodic_output(const struct device *dev, uin
 __syscall int tgpio_pin_read_ts_ec(const struct device *dev, uint32_t pin, uint64_t *timestamp,
 				    uint64_t *event_count);
 
-static inline int z_impl_tgpio_pin_read_ts_ec(const struct device *dev, uint32_t pin,
-					       uint64_t *timestamp, uint64_t *event_count)
-{
-	const struct tgpio_driver_api *api = (const struct tgpio_driver_api *)dev->api;
-
-	return api->read_ts_ec(dev, pin, timestamp, event_count);
-}
-
 /**
  * @}
  */
@@ -190,6 +143,7 @@ static inline int z_impl_tgpio_pin_read_ts_ec(const struct device *dev, uint32_t
 }
 #endif
 
+#include <zephyr/drivers/misc/timeaware_gpio/internal/timeaware_gpio_impl.h>
 #include <zephyr/syscalls/timeaware_gpio.h>
 
 #endif /* ZEPHYR_DRIVERS_MISC_TIMEAWARE_GPIO_TIMEAWARE_GPIO */
